add style module and change structure (move intarface to filter.page folder,  and separate interface to each own file)
