+++ src/pages/filter/components/button/button.interface.ts
@@ -1,0 +1,5 @@
+export interface ButtonProps {
+  text: string;
+  className: string;
+  onClick?: (event: React.MouseEvent<HTMLElement>) => void;
+}
