+++ src/pages/filter/components/button/button.module.scss
@@ -1,0 +1,25 @@
+:global {
+  .btn {
+    width: 107px;
+    height: 32px;
+    font-family: Roboto;
+    font-style: normal;
+    font-weight: normal;
+    font-size: 14px;
+    line-height: 16px;
+    border: none;
+    border-radius: 5px;
+    margin-right: 28px;
+    margin-top: 25px;
+  }
+
+  .btn__apply {
+    background-color: #902111;
+    color: #ffffff;
+  }
+
+  .btn__cancel {
+    background-color: #f7f7f6;
+    color: #000000;
+  }
+}
