+++ src/pages/filter/components/filter-body/bodyFilter.intarface.ts
@@ -1,0 +1,18 @@
+export interface BodyProps {
+  name: string;
+  id: string;
+  value: string;
+  lable: string;
+}
+
+export interface Data {
+  timeWolk: string;
+  moneyPerMonth: string;
+  timePerMonth: string;
+  securityLevel: string;
+  petSize: string;
+  easyToTrain: boolean;
+  family: boolean;
+  apartment: boolean;
+  allergy: boolean;
+}
