+++ src/pages/filter/components/input-radio/input-radio.module.scss
@@ -1,0 +1,47 @@
+:global {
+  .form_radio_group {
+    display: flex;
+    justify-content: space-between;
+    margin-top: 25px;
+  }
+
+  .form_radio_group input[type='radio'] {
+    display: none;
+  }
+
+  .form_radio_group label {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 74px;
+    height: 31px;
+    background: #f7f7f6;
+    border-radius: 15px;
+  }
+
+  .form_radio_group .form_radio_group-item:first-child label {
+    border-radius: 6px 0 0 6px;
+  }
+
+  .form_radio_group .form_radio_group-item:last-child label {
+    border-radius: 0 6px 6px 0;
+    border-right: 1px solid #999;
+  }
+
+  /* Checked */
+  .form_radio_group input[type='radio']:checked + label {
+    background: #902111;
+    color: #ffffff;
+  }
+
+  /* Hover */
+  .form_radio_group label:hover {
+    color: #666;
+  }
+
+  /* Disabled */
+  .form_radio_group input[type='radio']:disabled + label {
+    background: #efefef;
+    color: #666;
+  }
+}
