+++ src/pages/filter/components/input-switch/input-switch.component.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
-import { InputRadioProps } from '../../../comparison/interfaces/filter.interface';
+import { InputRadioProps } from '../../filter.interface';
+import style from './input-switch.module.scss';
 
 export const InputSwitch: React.FC<Partial<InputRadioProps>> = props => {
   return (
-    <div className="switch">
+    <div className={style.switch}>
       <input type="checkbox" name={props.name} ref={props.refAttribute} />
-      <span className="slider"></span>
+      <span className={style.slider}></span>
     </div>
   );
 };
