+++ src/pages/filter/components/overview-checkboxes/overview-checkboxes.component.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { InputSwitch } from '../input-switch/input-switch.component';
-import { InputRadioProps } from '../../../comparison/interfaces/filter.interface';
+import { InputRadioProps } from '../../filter.interface';
+import './overview-checkboxes.module.scss';
 
 export const OverviewCheckboxes: React.FC<Partial<InputRadioProps>> = props => {
   return (
