+++ src/pages/filter/components/slider-section/slider-section.module.scss
@@ -1,0 +1,4 @@
+.range__slider {
+  margin-top: 20px;
+  position: relative;
+}
