+++ src/pages/filter/components/slider/slider.module.scss
@@ -1,7 +1,3 @@
-.range-slider {
-  margin-top: 20px;
-  position: relative;
-}
 output {
   position: absolute;
   font-size: 14px;
