+++ src/pages/filter/components/timePerMonth-checkboxes/timePerMonth-checkboxes.component.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { InputRadio } from '../input-radio/input-radio.component';
-import { InputRadioProps } from '../../../comparison/interfaces/filter.interface';
+import { InputRadioProps } from '../../filter.interface';
 
 export const TimePerMonthCheckboxes: React.FC<Partial<InputRadioProps>> = props => {
   return (
