+++ src/pages/filter/filter.scss
@@ -6,114 +6,3 @@
   max-height: calc(100vh - 1px);
   overflow-y: auto;
 }
-
-.general {
-  margin-top: 35px;
-}
-
-.general label {
-  display: flex;
-  justify-content: space-between;
-  font-size: 14px;
-  line-height: 16px;
-  font-family: Roboto;
-  margin-top: 10px;
-}
-
-.switch {
-  position: relative;
-  width: 50px;
-  height: 34px;
-}
-
-.switch input {
-  opacity: 0;
-  width: 0;
-  height: 0;
-}
-
-.slider {
-  position: absolute;
-  cursor: pointer;
-  background-color: #f7f7f6;
-  -webkit-transition: 0.4s;
-  transition: 0.4s;
-  width: 50px;
-  height: 22px;
-  border: 1px solid #902111;
-  border-radius: 8px;
-}
-
-.slider:before {
-  position: absolute;
-  content: '';
-  height: 16px;
-  width: 16px;
-  left: 3px;
-  top: 2px;
-  border-radius: 10px;
-  -webkit-transition: 0.4s;
-  transition: 0.4s;
-  background-color: #902111;
-}
-
-input:checked + .slider {
-  background-color: #b5a79f;
-}
-
-input:focus + .slider {
-  box-shadow: 0 0 1px #b5a79f;
-}
-
-input:checked + .slider:before {
-  -webkit-transform: translateX(26px);
-  -ms-transform: translateX(26px);
-  transform: translateX(26px);
-  background-color: #1f1712;
-}
-
-.form_radio_group {
-  display: flex;
-  justify-content: space-between;
-  margin-top: 25px;
-}
-
-.form_radio_group input[type='radio'] {
-  display: none;
-}
-
-.form_radio_group label {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 74px;
-  height: 31px;
-  background: #f7f7f6;
-  border-radius: 15px;
-}
-
-.form_radio_group .form_radio_group-item:first-child label {
-  border-radius: 6px 0 0 6px;
-}
-
-.form_radio_group .form_radio_group-item:last-child label {
-  border-radius: 0 6px 6px 0;
-  border-right: 1px solid #999;
-}
-
-/* Checked */
-.form_radio_group input[type='radio']:checked + label {
-  background: #902111;
-  color: #ffffff;
-}
-
-/* Hover */
-.form_radio_group label:hover {
-  color: #666;
-}
-
-/* Disabled */
-.form_radio_group input[type='radio']:disabled + label {
-  background: #efefef;
-  color: #666;
-}
